Moved functions outside of ui
